Admin Tools
This small changeset, from over the weekend, adds some new admin tools
to the dashboard to centralize access to some of the more common places
admins have to go to run the system.

In particular, infrastructure monitoring charts have been added, with
plans for analytics and settings panes as well. This can be new fertile
ground for testing and proving out UI-side features on the dashboard
before they go live to end users.

Changes and enhancements
- [x] Add content for admin page at all
- [x] Add components for infrastructure charts and controls
- [ ] Add components for analytics charts
- [ ] Add pane for logged-in users
- [ ] Setup platform to report logged-in users
- [ ] Be able to execute RPC requests from Polymer
  - [ ] As admins (this should already work)
  - [ ] As supervisors / end users